% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{create_git_builder}
\alias{create_git_builder}
\title{Create Git Builder}
\usage{
create_git_builder(
  source,
  git_url,
  branch,
  credential,
  subdir,
  wait,
  build_trigger,
  url = ppm_url()
)
}
\arguments{
\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\description{
Create a new git builder for a git source
}
