% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{detect_distro}
\alias{detect_distro}
\title{Detect the PPM distribution matching this system}
\usage{
detect_distro()
}
\value{
A string containing the Package Manager distribution name
}
\description{
Attempts to detect the Package Manager binary distribution that
corresponds to the local operating system where R is running.
}
