% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_package_count}
\alias{get_package_count}
\title{Get package download count metrics}
\usage{
get_package_count(package_name, days = 30, url = ppm_url())
}
\arguments{
\item{package_name}{Package to get metrics for}

\item{days}{number of days to look back}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
Count of package downloads
}
\description{
Returns the total number of times a package has been downloaded
from the server in the past n days
}
\examples{
ggplot2_counts_90 <- get_package_count("ggplot2", days = 90)
dplyr_counts_90 <- get_package_count("dplyr", days = 90)
shiny_counts_90 <- get_package_count("shiny", days = 90)
}
