% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_package_count_history}
\alias{get_package_count_history}
\title{Get table of historic daily package counts}
\usage{
get_package_count_history(
  package_names,
  start_date = "2023-01-01",
  url = ppm_url()
)
}
\arguments{
\item{package_names}{character vector of package names to include}

\item{start_date}{starting date to search}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A data.frame listing downloads for each package by date
}
\description{
Get table of historic daily package counts
}
\examples{
package_daily_counts_90 <- get_package_count_history(c("ggplot2", "dplyr", "shiny"), start_date = today() - days(90))
}
