% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_builder_logs}
\alias{git_builder_logs}
\title{Retrieve logs from a Git package build}
\usage{
git_builder_logs(transaction_id, url = ppm_url())
}
\arguments{
\item{url}{}
}
\description{
Retrieve logs from a Git package build
}
