% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{list_distributions}
\alias{list_distributions}
\title{List binary distributions}
\usage{
list_distributions(url = ppm_url())
}
\arguments{
\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A data.frame of distribution names and a display name
}
\description{
Retrieve a list of available distributions for binary R packages from Package Manager.
}
\examples{
distros <- list_distributions()
}
