% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{list_packages}
\alias{list_packages}
\title{List packages}
\usage{
list_packages(
  source_name,
  repo_name,
  fields = c("name", "version"),
  url = ppm_url()
)
}
\arguments{
\item{source_name}{The source name}

\item{repo_name}{The repository name}

\item{fields}{Package properties to retrieve}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A tibble containing one row per package and one column per field
}
\description{
Specify either source_name or repoName.
}
