% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{list_repos}
\alias{list_repos}
\title{List all repositories on server}
\usage{
list_repos(url = ppm_url())
}
\arguments{
\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A tibble of repository names
}
\description{
List all repositories on server
}
