% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{list_snapshots}
\alias{list_snapshots}
\title{List all available snapshot dates for a repository}
\usage{
list_snapshots(repo_name, url = ppm_url())
}
\arguments{
\item{repo_name}{The repository name}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A tibble of snapshot dates
}
\description{
List all available snapshot dates for a repository
}
