% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{list_sources}
\alias{list_sources}
\title{List all sources}
\usage{
list_sources(url = ppm_url())
}
\arguments{
\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A tibble of source names
}
\description{
List all sources
}
