% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{P3M}
\alias{P3M}
\title{Default URL to Posit Public Package Manager}
\format{
An object of class \code{character} of length 1.
}
\usage{
P3M
}
\description{
Default URL to Posit Public Package Manager
}
\keyword{datasets}
