% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{ppm_get_repo_id}
\alias{ppm_get_repo_id}
\title{Get a repos ID from the repository name}
\usage{
ppm_get_repo_id(repo_name, url = ppm_url())
}
\arguments{
\item{repo_name}{The repository name}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\description{
Get a repos ID from the repository name
}
