% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{ppm_get_source_id}
\alias{ppm_get_source_id}
\title{Get a source ID from the source name}
\usage{
ppm_get_source_id(source_name, url = ppm_url())
}
\arguments{
\item{source_name}{The source name}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\description{
Retrieves the internal source ID for the given source name.
}
