% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmtools-package.R
\docType{package}
\name{ppmtools-package}
\alias{ppmtools}
\alias{ppmtools-package}
\title{ppmtools: Tools for Working With Posit Package Manager}
\description{
Tools to interact with the Posit Package Manager API and publish packages to local repositories.
}
\author{
\strong{Maintainer}: Joe Roberts \email{joe.roberts@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
