% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publish a package to Posit Package Manager}
\usage{
publish(source_name, url = ppm_url(), pkg = ".", args = NULL, binary = FALSE)
}
\arguments{
\item{source_name}{The Package Manager source to publish the package to}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}

\item{pkg}{The package}

\item{args}{Additional args passed to `pkgbuild::build`}

\item{binary}{Build and publish binary package instead of source?}
}
\description{
Publish a package to Posit Package Manager
}
