% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Status API}
\usage{
status(url = ppm_url())
}
\arguments{
\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\description{
Fetch various properties from the Package Manager /status endpoint
}
