% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{verify_config}
\alias{verify_config}
\title{Verify R is properly configured to use Posit Package Manager}
\usage{
verify_config(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, silence most messages.}
}
\value{
Invisibly returns any necessary config commands to be executed, otherwise NULL.
}
\description{
Verify R is properly configured to use Posit Package Manager
}
