% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure}
\alias{configure}
\title{Configure your R session to use Posit Package Manager}
\usage{
configure(url = ppm_url(), distro = NULL)
}
\arguments{
\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}

\item{distro}{PPM package distribution from ppmtools::list_distributions().  If NULL (default), attempt to auto-detect.}
}
\description{
Configure your R session to use Posit Package Manager
}
