% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{list_packages}
\alias{list_packages}
\title{List packages}
\usage{
list_packages(
  repo_name,
  source_name,
  search = "",
  fields = c("name", "version"),
  max_rows = 100,
  url = ppm_url()
)
}
\arguments{
\item{repo_name}{The repository name}

\item{source_name}{The source name}

\item{search}{A search string to perform a substring match against package names}

\item{fields}{Package properties to retrieve}

\item{max_rows}{Maximum number of packages to return.}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A tibble containing one row per package and one column per field
}
\description{
Specify either source_name or repo_name.
}
