% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{list_repos}
\alias{list_repos}
\title{List all repositories on server}
\usage{
list_repos(all = FALSE, url = ppm_url())
}
\arguments{
\item{all}{If TRUE include hidden repositories. Defaults to FALSE.}

\item{url}{The Package Manager URL.  Defaults to value of `options("ppm.url")`.}
}
\value{
A data.frame of repositories
}
\description{
List all repositories on server
}
